Add and remove tabs dynamically
Grouped controllers in package for easier management.
Also change concept URLs to need the concept name instead of their id because it was necessary to have the names for the tabs, so also we could improve the URLs to something more recognizable for the user.
Tabs still can't be closed.